Added Wishlist to product and added inWishlist method

+++ api/Modules/ShoppingCart/Traits/HasWishlist.php
@@ -17,14 +17,27 @@ trait HasWishlist{
     /*----------------------------------------------------
     * Methods
     --------------------------------------------------- */
+    public function inWishlist(Product $product)
+    {
+        return $this->products()->where('products.id', $product->id)->count();
+    }
+
     public function addToWishlist(Product $product)
     {
+        if($this->inWishlist($product)){
+            return $this;
+        }
+
         $this->products()->attach($product->id);
         return $this;
     }
 
     public function removeFromWishlist(Product $product)
     {
+        if (!$this->inWishlist($product)) {
+            return $this;
+        }
+
         $this->products()->detach($product->id);
         return $this;
     }
